Fit a raschdash model

Create an rdfit object from dimensional data.

new_rdfit(
  cohorts,
  groups,
  persons,
  person_groups,
  testlets,
  items,
  item_testlets,
  item_max_scores,
  K = 0,
  obs_group_cohorts,
  obs_groups,
  obs_group_items,
  obs_group_scores,
  obs_person_cohorts,
  obs_persons,
  obs_person_items,
  obs_person_scores,
  ...
)

rdfit(
  data,
  cohort = NULL,
  group,
  person,
  testlet,
  item,
  max_score,
  K = 0,
  obs_score,
  ...
)

Arguments

cohorts
vector of unique cohort identifiers

groups
vector of unique group identifiers

persons, person_groups
vectors of unique person identifiers (persons) and the identifiers of their corresponding groups (person_groups)

testlets
vector of unique testlet identifiers

items, item_testlets, item_max_scores
vectors of unique item identifiers (items) and their corresponding testlets (item_testlets) and maximum scores(max_scores)

K
maximum score of rating-scale items. Defaults to no rating scales.

obs_group_cohorts, obs_groups, obs_group_items, obs_group_scores
for the group-level observations, vectors of the corresponding cohorts (obs_group_cohorts), groups (obs_groups), items (obs_group_items), and observed scores (obs_group_scores)

obs_person_cohorts, obs_persons, obs_person_items, obs_person_scores
for the person-level observations, vectors of the corresponding cohorts (obs_person_cohorts), persons (obs_persons), items (obs_person_items), and observed scores (obs_person_scores)

...
additional parameters passed to rstan::sampling()

data
a data frame with columns for cohort, group, person, testlet, item, maximum possible item score, and observed score. For group-level observations, the person should be NA.

cohort
tidyr::tidyr_tidy_select() cohort column. Defaults to a generic All cohort.

group
tidyr::tidyr_tidy_select() group column

person
tidyr::tidyr_tidy_select() person column

testlet
tidyr::tidyr_tidy_select() testlet column

item
tidyr::tidyr_tidy_select() item column

max_score
tidyr::tidyr_tidy_select() maximum-score column

obs_score
tidyr::tidyr_tidy_select() observed score column

Value

An rdfit object, which has the following attributes.

stanfit

the rstan::stanfit() object for the fitted model. It contains samples of the following parameters.

logit-scale parameters

xi

group abilities

eta

person abilities

epsilon

testlet difficulties

delta

item difficulties

tau

rating-scale threshold offsets from overall item difficulty (if any rating-scale items are present)

lambda

standard deviation of person abilities. Used to transform the logit scale to the standard scale.

standard-scale parameters

group_ability

group abilities

person_ability

person abilities

testlet_difficulty

testlet difficulties

item_difficulty

item difficulties

thresholds

rating-scale threshold offsets from overall item difficulty (if any rating-scale items are present)

standard-scale parameters under their prior distribution given the posterior distribution of the hyper-parameters

prior_group_ability

prior_person_ability

prior_testlet_difficulty

prior_item_difficulty

posterior predictive distribution

y_rep

predicted score for each observed combination of group, person, and item

log likelihoods

log_lik

log likelihood of observed scores

log_lik_rep

log likelihood of predicted scores

loo

a loo::loo() object for stanfit

Details

The rdfit function is intended to be the primary method for users to create rdfit objects. Most R users work with so-called dimensional data in a single data frame. In the case that the data are alread stored in a normalised form, however, the raw new_rdfit constructor will work. Be aware that new_rdfit does not validate whether the data are properly normalised, and Stan will throw rather cryptic error message if they are not.

Functions

new_rdfit: Create an rdfit object from normalised data.