Fit a raschdash model

Create rdfit objects from observed data.

rdfit(
  data,
  cohort = "All",
  group = "All",
  person,
  testlet = "All",
  item,
  max_score,
  obs_score,
  k = 1,
  ...
)

new_rdfit(
  cohorts,
  groups,
  persons,
  person_groups,
  testlets,
  items,
  item_testlets,
  item_max_scores,
  obs_group_cohorts,
  obs_groups,
  obs_group_items,
  obs_group_scores,
  obs_person_cohorts,
  obs_persons,
  obs_person_items,
  obs_person_scores,
  k = 1,
  pars = c("standard", "logit", "hyper", "all"),
  ...
)

Arguments

data
a data frame with columns for cohort, group, person, testlet, item, maximum possible item score, and observed score. For group-level observations, the person should be NA.

cohort
<tidyr::tidyr_tidy_select> cohort column. Defaults to a generic All cohort.

group
<tidyr::tidyr_tidy_select> group column. Defaults to a generic All group.

person
<tidyr::tidyr_tidy_select> person column

testlet
<tidyr::tidyr_tidy_select> testlet column. Defaults to a generic All testlet.

item
<tidyr::tidyr_tidy_select> item column

max_score
<tidyr::tidyr_tidy_select> maximum-score column

obs_score
<tidyr::tidyr_tidy_select> observed score column

k
maximum score of rating-scale items. Defaults to no rating scales.

...
additional parameters passed to rstan::sampling()

cohorts
vector of unique cohort identifiers

groups
vector of unique group identifiers

persons, person_groups
vectors of unique person identifiers (persons) and the identifiers of their corresponding groups (person_groups)

testlets
vector of unique testlet identifiers

items, item_testlets, item_max_scores
vectors of unique item identifiers (items) and their corresponding testlets (item_testlets) and maximum scores(max_scores)

obs_group_cohorts, obs_groups, obs_group_items, obs_group_scores
for the group-level observations, vectors of the corresponding cohorts (obs_group_cohorts), groups (obs_groups), items (obs_group_items), and observed scores (obs_group_scores)

obs_person_cohorts, obs_persons, obs_person_items, obs_person_scores
for the person-level observations, vectors of the corresponding cohorts (obs_person_cohorts), persons (obs_persons), items (obs_person_items), and observed scores (obs_person_scores)

pars
which parameters to save from the Stan model. Defaults to standard, which only saves parameters on the standard scale, but may also be logit to add logit-scale parameters, or hyper to add logit-scale parameters and hyper-parameters. For debugging, the all option adds other raw parameters specific to the Stan implementation.

Value

An object of class rdfit with the following attributes.

data

The input data as passed to Stan.

stanfit

The rstan::stanfit object for the fitted model. It contains samples of the following parameters.

standard-scale parameters

group_ability: group abilities

person_ability: person abilities

testlet_difficulty: testlet difficulties

item_difficulty: item difficulties

thresholds: rating-scale threshold offsets from overall item difficulty (if any rating-scale items are present)

sigma: standard deviation of person abilities on the logit scale. Used to convert between the logit and standard scales.

posterior predictive distribution

y_rep: predicted score for each observed combination of group, person, and item

prior_group_ability (sampled under posterior hyper-parameters)

prior_person_ability (sampled under posterior hyper-parameters)

prior_testlet_difficulty (sampled under posterior hyper-parameters)

prior_item_difficulty (sampled under posterior hyper-parameters)

log likelihoods

log_lik: log likelihood of observed scores

log_lik_rep: log likelihood of predicted scores

logit-scale parameters (when pars = "logit" or pars = "hyper")

xi: group abilities

eta: person abilities

epsilon: testlet difficulties

delta: item difficulties

tau: rating-scale threshold offsets from overall item difficulty (if any rating-scale items are present)

hyper-parameters (when pars = "hyper")

nu: mean difficulty

psi: scale of group abilities

phi: scale of person abilities (relative to group)

theta_epsilon: scale of testlet difficulties

theta_upsilon: scale of item difficulties (relative to testlet)

loo

A loo::loo object for stanfit.

Details

The rdfit() function is intended to be the primary method for users to create rdfit objects. Most R users work with denormalised dimensional data in a single data frame. In the case that the data are already stored in a normalised structure, however, the raw new_rdfit() constructor will work. The new_rdfit() constructor does not validate whether the data are properly normalised, however, and Stan will throw sometimes cryptic error messages if they are not.

Functions

rdfit: Create an rdfit object from dimensional data.

new_rdfit: Create an rdfit object from normalised data.